+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 - The RBAC ServiceAccount and RoleBinding are now built with the operator-rs `v2::rbac`
   functions and carry the full set of recommended labels ([#846]).
 - Bump stackable-operator to 0.114.0 ([#855]).
+- The reconciler now applies resources and derives the cluster status in discrete
+  apply and update_status steps ([#856]).
 - All product containers now run with `securityContext.runAsNonRoot` set to `true` to improve security ([#860]).
 
 ### Fixed
@@ -22,6 +24,7 @@ All notable changes to this project will be documented in this file.
 [#841]: https://github.com/stackabletech/druid-operator/pull/841
 [#846]: https://github.com/stackabletech/druid-operator/pull/846
 [#855]: https://github.com/stackabletech/druid-operator/pull/855
+[#856]: https://github.com/stackabletech/druid-operator/pull/856
 [#860]: https://github.com/stackabletech/druid-operator/pull/860
 
 ## [26.7.0] - 2026-07-21